# openapi-spec.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/grc_policy_query": {
            "post": {
                "tags": [
                    "Docs"
                ],
                "summary": "Retrieval Generate",
                "operationId": "retrieval_generate_api_grc_policy_query_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Auth",
                            "type": "string"
                        },
                        "name": "auth",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/grc_queryPayLoad"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/simple_query": {
            "post": {
                "tags": [
                    "Docs"
                ],
                "summary": "Retrieval Generate",
                "operationId": "retrieval_generate_api_simple_query_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Auth",
                            "type": "string"
                        },
                        "name": "auth",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/simpleTprmQueryPayLoad"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
...
            "grcPolicy": {
                "title": "grcPolicy",
                "required": [
                    "policy_name",
                    "policy_code",
                    "controls",
                    "evidences"
                ],
                "type": "object",
                "properties": {
                    "policy_name": {
                        "title": "Policy Name",
                        "type": "string"
                    },
                    "policy_code": {
                        "title": "Policy Code",
                        "type": "string"
                    },
                    "controls": {
                        "title": "Controls",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/grcPolicyControl"
                        }
                    },
                    "evidences": {
                        "title": "Evidences",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/grcPolicyEvidence"
                        }
                    }
                }
            },
            "grcPolicyControl": {
                "title": "grcPolicyControl",
                "required": [
                    "control_code",
                    "control_quetion"
                ],
                "type": "object",
                "properties": {
                    "control_code": {
                        "title": "Control Code",
                        "type": "string"
                    },
                    "control_quetion": {
                        "title": "Control Quetion",
                        "type": "string"
                    }
                }
            },
            "grcPolicyEvidence": {
                "title": "grcPolicyEvidence",
                "required": [
                    "evidence_code",
                    "evidence_artifact"
                ],
                "type": "object",
                "properties": {
                    "evidence_code": {
                        "title": "Evidence Code",
                        "type": "string"
                    },
                    "evidence_artifact": {
                        "title": "Evidence Artifact",
                        "type": "string"
                    }
                }
            },
...
            "grc_queryPayLoad": {
                "title": "grc_queryPayLoad",
                "required": [
                    "user_id",
                    "doc_type",
                    "policies"
                ],
                "type": "object",
                "properties": {
                    "user_id": {
                        "title": "User Id",
                        "type": "string"
                    },
                    "doc_type": {
                        "title": "Doc Type",
                        "type": "string"
                    },
                    "policies": {
                        "title": "Policies",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/grcPolicy"
                        }
                    }
                }
            },
...
            "simpleTprmQueryPayLoad": {
                "title": "simpleTprmQueryPayLoad",
                "required": [
                    "question"
                ],
                "type": "object",
                "properties": {
                    "question": {
                        "title": "Question",
                        "type": "string"
                    },
                    "customer_id": {
                        "title": "Customer Id",
                        "type": "string"
                    },
                    "context_status": {
                        "title": "Context Status",
                        "type": "boolean",
                        "default": true
                    }
                }
            },
...
        }
    }
}